third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-audioparam-interface/audioparam-default-value.window-expected.txt

This is a testharness.js-based test.
[FAIL] AudioParam's defaultValue
  assert_equals: AudioParam's defaultValue is not correct. expected -1 but got 1
Harness: the test ran to completion.
